Make test DSL generic over the type of certificate
